[mod_dptools:playback] fix audio speed change to preserve pitch
When  a  file  is  being played  with  the  playback
application  and  the  speed  is  changed  with  eg.
uuid_fileman  speed:+1  the audio  pitch  increases.
This commit tries to preserve the pitch even when the
speed is changed.

How?
+ Use longer sample chunks (where original freq is preserved)
+ Search for the best cut position using cross-correlation
+ Crossfade adjacent chunks to each other

This is  basically a copy  of my unreviewed  PR from
the old system
https://freeswitch.org/stash/projects/FS/repos/freeswitch/pull-requests/1707/commits/b8b98bbf310c6d8f138f68af7815d0938c440a13