[mod_sofia] Randomize OPTIONS Ping interval
In FS-6400, the attempt was made to randomize OPTIONS
packets to be sent at a random interval.  The same random
interval is applied to all endpoints so this doesn't work.

Furthermore, rounding withing the code will ultimately
make the ping times converge over time.

This commmit will apply the random interval only during
registration of the endpoint.  All subsequent pings will
be incremented with the actual value of ping-mean-interval.

Fixes #1132, Fixes #1133