[mod_sofia] Call srand() only once per thread
In randomization of numbers for ping intervals,
srand was called each time the randomization
function was called.  This is incorrect as all
the registrations were getting the same value
instead of a random one.

We move srand() to the beginning of the worker
thread launch in order to generate random numbers
correctly.

Fixes #1133